server/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
gevent==24.2.1
